# Improved performance of forms.widgets.Media.merge()