As highlighted in https://github.com/django/django/pull/13874 `OrderedSet` can be a little "slow". Now that we can rely on dictionaries to be ordered the quickest way to get an ordered list of unique items is `list(dict.fromkeys(iterable))`. 

While `OrderedSet` is not used that much in Django's codebase, one example is in widget `Media`. My [benchmark](https://gist.github.com/smithdc1/f05610c6c28954e88c345d67e793405c) shows this patch is 13% quicker than the current `OrderedSet` approach. 

@felixxm -- Does this change need a ticket? -- I'd like to review the other uses for `OrderedSet` but to review them properly and to write benchmarks will take time. I'm also conscious of the upcoming feature freeze -- so there's an option of considering this as a standalone change, or we could wait for a more comprehensive review.