@felixxm 2021-01-26
@smithdc1 Thanks for this patch :+1: however I'm not sure if it's worth readability, there is a small performance improvement but both methods are really fast :thinking: :

- Python 3.10:

  - before: `Mean +- std dev: 7.72 us +- 0.64 us`
  - after: `Mean +- std dev: 6.73 us +- 0.39 us`

- Python 3.8:

  - before: `Mean +- std dev: 6.64 us +- 0.41 us`
  - after: `Mean +- std dev: 6.02 us +- 0.32 us`